added GroupLpasInterface.php and moved interface to CombinedLpa.php
