+++ service-front/app/src/Common/src/Service/Lpa/ServiceInterfaces/GroupLpasInterface.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Common\Service\Lpa\ServiceInterfaces;
+
+use Common\Entity\Person;
+
+interface GroupLpasInterface
+{
+    public function getDonor(): Person;
+
+    public function getCaseSubtype(): string;
+}
