+++ service-front/app/src/Common/src/Service/Lpa/ServiceInterfaces/SortLpasInterface.php
@@ -2,12 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Common\Service\Lpa;
+namespace Common\Service\Lpa\ServiceInterfaces;
 
 use Common\Entity\Person;
-use EventSauce\ObjectHydrator\MapperSettings;
 
-#[MapperSettings(serializePublicMethods: false)]
 interface SortLpasInterface
 {
     public function getDonor(): Person;
